Import: avoid repeatedly (de)init provider wrapper
The provider wrapper is used as an immutable object during importing.
There is no need to repeatedly launch and kill it in each iteraton on
services.

This will brings big performance improvement on some of the provider
, which will do some heavy-lift on initialization. For example, azurerm
provider will do auto registering resource providers on start up, which
will take about 10 seconds. For commands that specify a `-r "*"`,
terraformer will iterate each supported services, and each one will
take 10 seconds, which ends up to be a fairly long interval.