Fix missing output variables when compact output
When user specify `-C`, current code doesn't record the referenced
attributes in the output config.

This PR addresses that issue by adding an additional check on whether
we are in compact mode. If yes, then it means the `resources` under
iteration belongs to different services, and we will only record the
output for current resource when the connection's target service is
the same as this resource's service type.