Upgrade Keycloak provider to version 4.0.1 (#1515)
* feat(keycloak)!: upgrade provider version to 4.0.1

Introduce `base_path` and `red_hat_sso` provider arguments. These can be
overriden with `KEYCLOAK_BASE_PATH` and `RED_HAT_SSO` environment
variables. Users of the legacy version of Keycloak will need to set the
base path to `/auth`.

The `OpenidClientWithGenericClientProtocolMappers` has been renamed to
`OpenidClientWithGenericProtocolMappers`.

The `GetGenericClientProtocolMappers` has been renamed to
`GetGenericProtocolMappers`.

Add support for `oidc-audience-resolve-mapper` and
`oidc-script-based-protocol-mapper` types.

Remove support for `oidc-hardcoded-group-mapper` as there is no such
protocol mapper in Keycloak.

Set some arbitrary string in the User-Agent request header.

Refs: #1229

* feat(keycloak): use realm instead of id

Keycloak 19 generates a UUID for the ID when a realm is created on the
web UI. Most requests with that internal ID results in a 404. The only
resource where we need to specify the internal ID is in the
`custom_user_federation`'s `parent_id` attribute.

Refs: #790

* docs: correct spelling and typos